Refactor controllers, services and repositories for a dealing with errors; finished amenities;

+++ src/services/buildingService.ts
@@ -1,4 +1,6 @@
+import { httpStatusEnum } from "../enums/httpStatusEnum.js";
 import { RolesEnum } from "../enums/rolesEnum.js";
+import Building from "../models/Building.js";
 import User from "../models/User.js";
 import buildingRepository from "../repositories/buildingRepository.js";
 export default class buildingService {
@@ -13,9 +15,17 @@ export default class buildingService {
 
         try {
 
-            await buildingService.validate(building);
+            let validation = await this.#validation(building);
 
-            return await buildingRepository.insert(building);
+            if(!validation.length) {
+
+                return await buildingRepository.insert(building);
+            }
+
+            throw {
+                status: httpStatusEnum.NOT_FOUND,
+                messages: validation
+            }
 
         } catch (err: any) {
 
@@ -28,9 +38,17 @@ export default class buildingService {
 
         try {
 
-           await this.validate(building);
+            let validation = await this.#validation(building);
 
-            return await buildingRepository.update(building);
+            if(!validation.length) {
+
+                return await buildingRepository.update(building);
+            }
+
+            throw {
+                status: httpStatusEnum.NOT_FOUND,
+                messages: validation
+            }
 
         } catch (err: any) {
 
@@ -41,30 +59,76 @@ export default class buildingService {
 
     static async delete(id: any) {
 
+        let building = await Building.findOne({
+            where: {
+                id
+            }
+        });        
+
+        if (!building) {            
+
+            throw {
+                status: httpStatusEnum.NOT_FOUND,
+                messages: "Building not found"
+            }
+
+        }
+
         return await buildingRepository.delete(id);
 
     }
 
-    static async validate(building:any) {        
+    static async #validation(building:any) {
+
+        let messages = [];
+
+        let syndic = await this.#syndicExists(building.syndic_id);
+
+        if(!syndic) {
+
+            messages.push("Syndic not found");
+
+            return messages;
+        }
+
+        let isSyndic = await this.#isSyndic(syndic.role_id);
+
+        if(!isSyndic) {
+
+            messages.push("User is not a syndic");
+
+        }
+
+        return messages;
+
+    }
+
+    static async #syndicExists(syndicId:any) {        
 
         let user = await User.findOne({
             where: {
-                id: building.syndic_id
+                id: syndicId
             }
         });
 
         if (!user) {
 
-            throw "User not found";
+            return false;
 
         }
+        
+        return user.dataValues;
 
-        let { role_id } = user.dataValues;
+    }
 
-        if (role_id !== RolesEnum.SYNDIC) {
+    static async #isSyndic(roleId:number) {
+        
+        if (roleId !== RolesEnum.SYNDIC && roleId !== RolesEnum.ADMIN) {
 
-            throw "User is not a syndic";
+            return false;
         }
+
+        return true;
 
     }
